basic tcp communication with writing /reading to fpga PLB_LL_FIFO_REG_CTRL works
git-svn-id: file:///afs/psi.ch/project/sls_det_software/svn/slsDetectorSoftware@441 951219d9-93cf-4727-9268-0efd64621fa3

+++ slsDetectorSoftware/eigerDetectorServer/slsDetector_firmware.h
@@ -17,7 +17,8 @@
 
 //memory
 int 		mapCSP0(void);
-u_int32_t 	bus_w(u_int32_t offset, u_int32_t data);
+u_int32_t 	bus_w(u_int32_t offset, u_int8_t data);
+//u_int32_t 	bus_w(u_int32_t offset, u_int32_t data);
 u_int32_t 	bus_r(u_int32_t offset);
